Put Avatar and tooltip in a separate component
